# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Daemon Host Panel
  version: 0.1.0
paths:
...
  /api/participant/name:
    put:
      tags:
      - participant
      summary: Rename Participant
      description: Rename a registered participant. Returns 400 if not yet registered.
      operationId: rename_participant_api_participant_name_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameRequest'
        required: true
      responses:
        '204':
          description: Successful Response
      x-feature: identity
  /api/participant/avatar:
    post:
      tags:
      - participant
      summary: Refresh Avatar Endpoint
      description: Re-roll avatar (conference mode only).
      operationId: refresh_avatar_endpoint_api_participant_avatar_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarResponse'
      x-feature: identity
  /api/participant/location:
    post:
      tags:
      - participant
      summary: Set Location
      description: Store participant city/timezone.
      operationId: set_location_api_participant_location_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: identity
...
components:
  schemas:
...
    AvatarRequest:
      properties:
        rejected:
          items:
            type: string
          type: array
          title: Rejected
          default: []
      type: object
      title: AvatarRequest
    AvatarResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        avatar:
          type: string
          title: Avatar
      type: object
      required:
      - avatar
      title: AvatarResponse
...
    LocationRequest:
      properties:
        location:
          type: string
          title: Location
      type: object
      required:
      - location
      title: LocationRequest
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    RenameRequest:
      properties:
        name:
          type: string
          title: Name
      type: object
      required:
      - name
      title: RenameRequest
...